requirements.txt
atlassian-python-api==3.40.0